backend/requirements.txt
Flask==2.3.2
python-dotenv==1.0.0
requests==2.31.0
gunicorn==20.1.0
# Optional (only if you use Vertex AI)
# google-cloud-aiplatform==1.30.0
